refactor: report dto 패키지 분리
