+++ src/main/java/ssu/eatssu/domain/report/dto/response/ReportTypeInformation.java
@@ -1,4 +1,4 @@
-package ssu.eatssu.domain.report.dto;
+package ssu.eatssu.domain.report.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
