+++ src/main/java/ssu/eatssu/domain/report/presentation/ReportController.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import ssu.eatssu.domain.auth.security.CustomUserDetails;
-import ssu.eatssu.domain.report.dto.ReportCreateRequest;
-import ssu.eatssu.domain.report.dto.ReportTypeList;
+import ssu.eatssu.domain.report.dto.request.ReportCreateRequest;
+import ssu.eatssu.domain.report.dto.response.ReportTypeList;
 import ssu.eatssu.domain.report.presentation.docs.ReportControllerDocs;
 import ssu.eatssu.domain.report.service.ReportService;
 import ssu.eatssu.domain.review.entity.Report;
