+++ src/main/java/ssu/eatssu/domain/report/service/ReportService.java
@@ -5,8 +5,8 @@ import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ssu.eatssu.domain.auth.security.CustomUserDetails;
-import ssu.eatssu.domain.report.dto.ReportCreateRequest;
-import ssu.eatssu.domain.report.dto.ReportTypeList;
+import ssu.eatssu.domain.report.dto.request.ReportCreateRequest;
+import ssu.eatssu.domain.report.dto.response.ReportTypeList;
 import ssu.eatssu.domain.report.entity.ReportStatus;
 import ssu.eatssu.domain.report.repository.ReportRepository;
 import ssu.eatssu.domain.review.entity.Report;
